remove: autoComplete.js 삭제 및 코드 통합 [#26]
