+++ views/index.hbs
@@ -89,7 +89,6 @@
     </div>
   </body>
   <script type='module' src='/javascript/index.js'></script>
-  <script type='module' src='/javascript/autoComplete.js'></script>
   <script src='/javascript/spell/spellCheck.js'></script>
   <script src='/javascript/spell/popup.js'></script>
   <script src='/javascript/longSentence/longSentence.js'></script>
